pyproject.toml
[tool.poetry]
name = "project-generator"
version = "0.1.0"
description = "A configurable Python project generator supporting multiple templates and tools"
authors = ["Rory McGuinness <rjmcguinness@gmail.com>"]
readme = "README.md"
packages = [{include = "project_generator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
click = "^8.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.0.0"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
project-generator = "project_generator.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
